Moved Configuration to appropriate folder

+++ src/jdz/RTGen/AppConfig.java
@@ -3,10 +3,10 @@ package jdz.RTGen;
 
 import java.util.Random;
 
-import jdz.RTGen.dataType.Config;
+import jdz.RTGen.configuration.Config;
 
 public class AppConfig extends Config {
-	public static long MAP_SEED = new Random().nextInt();
+	public static int MAP_SEED = new Random().nextInt();
 	
 	public static int MAP_HEIGHT = 256;
